Fix saving attacks
Fixes an issue where a copy of song-level attacks was added to the
step-level attacks when saving a song that had attacks at both the song
level and the step level. It would be cleaner to not merge song and step
attacks, but maybe there are some charts that rely on that.