Fix global offset applied at song loading
Fixes the music being not synced across machines when
1. using sync start,
2. the machines have different global offsets, and
3. the global offsets were changed after the songs were loaded for the
   first time (i.e. not from the cache).

The song cache must be rebuilt after applying this patch.

Song::ReCalculateStepStatsAndLastSecond used the current value of the
global offset when setting the first second and last second of the song
during song loading. On subsequent launches, the values are loaded from
the song cache. The values are not updated when the global offset is
changed, so they will be wrong if the offset is modified afterwards.

The first second and last second only affect the delay before the music
starts playing and before it stops, so the difference isn't a problem in
normal use. But with sync start, the music will be off sync across the
machines.

Fixed by storing the first second and last second without the global
offset in Song objects and in the song cache.

I tested this patch by adding a 1000 ms audio delay and -1000 ms global
offset on one machine, and 0 ms delay and 0 ms offset on another. When
starting a song with sync start, the music was in sync across the
machines